requirements.txt
torch==2.1.0
torchvision==0.15.0
pandas==1.5.3
matplotlib==3.7.0
scikit-learn==1.2.2
Pillow==9.4.0
tqdm==4.64.1
